requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
black==20.8b1
blessings==1.7
bpython==0.21
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
curtsies==0.3.5
cwcwidth==0.1.4
Django==3.1.7
flake8==3.8.4
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
python-decouple==3.4
python-telegram-bot==13.3
pytz==2021.1
pyxdg==0.27
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
